See description
Change the import script architecture.
There are some reasons for this:

* New architecture improve understanding how import works.
* Code reading become easier (in some places).
* Now configuring import process become easier.
* The import script become faster, bc there are no extra foreach cycles now.
