+++ README.md
@@ -1,5 +1,9 @@
 ### Известные баги:
-* Скрипт по импорту данных не работает, ругается на слишком большое использование оперативной памяти.
-* Не работает прогрессбар по неизвестной причине, чинить это не буду, вместо этого сделаю output.
+* ~~Скрипт по импорту данных не работает, ругается на слишком большое использование оперативной памяти.~~
+* ~~Не работает прогрессбар по неизвестной причине, чинить это не буду, вместо этого сделаю output.~~  
+  
+Проблема со слишком большим использованием памяти решена путем изменения архитектуры всего скрипта и добавлением чтения по чанкам, 
+так же после изменения архитектуры начал показываться прогрессбар, но он почему-то не заполняется.  
+
 * Добрая часть данных из-за существующих ограничений не импортируется или импортируется не правильно.
 * На этапе импорта prev_license в таблице licenses запрос, который осуществляет поиск лицензии по связке номера серии и вида лицензии, иногда возвращает несколько значений
