+++ app/Console/Commands/ExcelImport.php
@@ -3,16 +3,7 @@
 namespace App\Console\Commands;
 
 use Illuminate\Console\Command;
-use App\Imports\FederalDistrictImport;
-use App\Imports\RfSubjectImport;
-use App\Imports\DevelopmentDegreeImport;
-use App\Imports\FederalAuthorityImport;
-use App\Imports\FieldCompositionImport;
-use App\Imports\FieldImport;
-use App\Imports\FieldPositionImport;
-use App\Imports\LicenseAreaImport;
-use App\Imports\LicenseImport;
-use App\Imports\SubsoilUserImport;
+use App\Imports\Import;
 
 class ExcelImport extends Command
 {
@@ -38,16 +29,7 @@ class ExcelImport extends Command
     public function handle()
     {
 		$this->output->title('Starting import');
-		(new FederalDistrictImport)->withOutput($this->output)->import('data.xlsx');
-		(new RfSubjectImport)->withOutput($this->output)->import('data.xlsx');
-		(new DevelopmentDegreeImport)->withOutput($this->output)->import('data.xlsx');
-		(new FieldImport)->withOutput($this->output)->import('data.xlsx');
-		(new FieldPositionImport)->withOutput($this->output)->import('data.xlsx');
-		(new SubsoilUserImport)->withOutput($this->output)->import('data.xlsx');
-		(new LicenseAreaImport)->withOutput($this->output)->import('data.xlsx');
-		(new FederalAuthorityImport)->withOutput($this->output)->import('data.xlsx');
-		(new LicenseImport)->withOutput($this->output)->import('data.xlsx');
-		(new FieldCompositionImport)->withOutput($this->output)->import('data.xlsx');
+		(new Import)->withOutput($this->output)->import('data.xlsx');
 		$this->output->success('Import successful');
         return 0;
     }
